QuillEditor component rendering Quill text editor
