+++ public/index.html
@@ -1,76 +1,96 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta name="theme-color" content="#000000">
-    <!--
-      manifest.json provides metadata used when your web app is added to the
-      homescreen on Android. See https://developers.google.com/web/fundamentals/engage-and-retain/web-app-manifest/
-    -->
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
-    <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+  <meta name="theme-color" content="#000000">
+  <!--
+  manifest.json provides metadata used when your web app is added to the
+  homescreen on Android. See https://developers.google.com/web/fundamentals/engage-and-retain/web-app-manifest/
+-->
+<link rel="manifest" href="%PUBLIC_URL%/manifest.json">
+<link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
 
-    <!-- need all of this below included for Summernote text editor to work? -->
-    <!-- include libraries(jQuery, bootstrap) -->
-    <!-- <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script> -->
-    <!-- <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.3.5/css/bootstrap.min.css" /> -->
-    <!-- <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.3.5/js/bootstrap.min.js"></script> -->
+<!-- need all of this below included for Summernote text editor to work? -->
+<!-- include libraries(jQuery, bootstrap) -->
+<!-- <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script> -->
+<!-- <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.3.5/css/bootstrap.min.css" /> -->
+<!-- <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.3.5/js/bootstrap.min.js"></script> -->
 
-    <!-- include summernote css/js-->
-    <!-- <link href="summernote.css" rel="stylesheet"> -->
-    <!-- <script src="summernote.js"></script> -->
-    <!-- need all of this above included for Summernote text editor to work? -->
+<!-- include summernote css/js-->
+<!-- <link href="summernote.css" rel="stylesheet"> -->
+<!-- <script src="summernote.js"></script> -->
+<!-- need all of this above included for Summernote text editor to work? -->
 
-    <!-- index.hmtl in Graphite Docs uses summernote-lite: -->
-    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"></script>
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/summernote/0.8.9/summernote-lite.css" rel="stylesheet">
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/summernote/0.8.9/summernote-lite.js"></script>
+<!-- index.hmtl in Graphite Docs uses summernote-lite: -->
+<script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"></script>
+<link href="https://cdnjs.cloudflare.com/ajax/libs/summernote/0.8.9/summernote-lite.css" rel="stylesheet">
+<script src="https://cdnjs.cloudflare.com/ajax/libs/summernote/0.8.9/summernote-lite.js"></script>
 
 
-    <!-- adding below for Quill text editor: -->
-    <!-- Main Quill library -->
-    <script src="//cdn.quilljs.com/1.0.0/quill.js"></script>
-    <script src="//cdn.quilljs.com/1.0.0/quill.min.js"></script>
+<!-- adding below for ReactQuill text editor: -->
+<script src="https://cdn.quilljs.com/1.1.9/quill.js"></script>
+<link href="https://cdn.quilljs.com/1.1.9/quill.snow.css" rel="stylesheet">
 
-    <!-- Theme included stylesheets -->
-    <link href="//cdn.quilljs.com/1.0.0/quill.snow.css" rel="stylesheet">
-    <link href="//cdn.quilljs.com/1.0.0/quill.bubble.css" rel="stylesheet">
+<!-- Core build with no theme, formatting, non-essential modules -->
+<!-- <script src="//cdn.quilljs.com/1.2.3/quill.core.js"></script> -->
+<!-- <link href="//cdn.quilljs.com/1.2.3/quill.core.css" rel="stylesheet"> -->
 
-    <!-- Core build with no theme, formatting, non-essential modules -->
-    <link href="//cdn.quilljs.com/1.0.0/quill.core.css" rel="stylesheet">
-    <script src="//cdn.quilljs.com/1.0.0/quill.core.js"></script>
-    
-    <!-- Include the Quill library -->
-    <script src="https://cdn.quilljs.com/1.3.6/quill.js"></script>
-    <!-- Include stylesheet -->
-    <link href="https://cdn.quilljs.com/1.3.6/quill.snow.css" rel="stylesheet"></link>
+<!-- Main Quill library -->
+<!-- <script src="//cdn.quilljs.com/1.2.3/quill.js"></script> -->
+<!-- <script src="//cdn.quilljs.com/1.2.3/quill.min.js"></script> -->
 
-    <!--
-      Notice the use of %PUBLIC_URL% in the tags above.
-      It will be replaced with the URL of the `public` folder during the build.
-      Only files inside the `public` folder can be referenced from the HTML.
+<!-- Theme included stylesheets -->
+<!-- <link href="//cdn.quilljs.com/1.2.3/quill.snow.css" rel="stylesheet"> -->
+<!-- <link href="//cdn.quilljs.com/1.2.3/quill.bubble.css" rel="stylesheet"> -->
 
-      Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
-      work correctly both with client-side routing and a non-root public URL.
-      Learn how to configure a non-root public URL by running `npm run build`.
-    -->
-    <title>React TextEdit App</title>
-  </head>
-  <body>
-    <noscript>
-      You need to enable JavaScript to run this app.
-    </noscript>
-    <div id="root"></div>
-    <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
+<!-- <link rel="stylesheet" href="//cdn.quilljs.com/1.2.6/quill.snow.css"> -->
+<!-- <link href="https://cdn.quilljs.com/1.3.6/quill.snow.css" rel="stylesheet"> -->
 
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
 
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
-    -->
-  </body>
+<!-- adding below for Quill text editor: -->
+<!-- Main Quill library -->
+<!-- <script src="//cdn.quilljs.com/1.0.0/quill.js"></script> -->
+<!-- <script src="//cdn.quilljs.com/1.0.0/quill.min.js"></script> -->
+
+<!-- Theme included stylesheets -->
+<!-- <link href="//cdn.quilljs.com/1.0.0/quill.snow.css" rel="stylesheet"> -->
+<!-- <link href="//cdn.quilljs.com/1.0.0/quill.bubble.css" rel="stylesheet"> -->
+
+<!-- Core build with no theme, formatting, non-essential modules -->
+<!-- <link href="//cdn.quilljs.com/1.0.0/quill.core.css" rel="stylesheet"> -->
+<!-- <script src="//cdn.quilljs.com/1.0.0/quill.core.js"></script> -->
+
+<!-- Include the Quill library -->
+<!-- <script src="https://cdn.quilljs.com/1.3.6/quill.js"></script> -->
+<!-- Include stylesheet -->
+<!-- <link href="https://cdn.quilljs.com/1.3.6/quill.snow.css" rel="stylesheet"></link> -->
+
+<!--
+Notice the use of %PUBLIC_URL% in the tags above.
+It will be replaced with the URL of the `public` folder during the build.
+Only files inside the `public` folder can be referenced from the HTML.
+
+Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
+work correctly both with client-side routing and a non-root public URL.
+Learn how to configure a non-root public URL by running `npm run build`.
+-->
+<title>React TextEdit App</title>
+</head>
+<body>
+  <noscript>
+    You need to enable JavaScript to run this app.
+  </noscript>
+  <div id="root"></div>
+  <!--
+  This HTML file is a template.
+  If you open it directly in the browser, you will see an empty page.
+
+  You can add webfonts, meta tags, or analytics to this file.
+  The build step will place the bundled scripts into the <body> tag.
+
+  To begin the development, run `npm start` or `yarn start`.
+  To create a production bundle, use `npm run build` or `yarn build`.
+-->
+</body>
 </html>
